Move spawnDrops()
